Fixed eligible / non-eligible players
Players must be in zone, not afk and PVP flagged
Winning side (elig : 1x rewards, non-elig 0.5x rewards)
Losing side (elig : 0.5x rewards, non-elig 0.25x rewards)
Eligible winner gets bag

+++ WorldServer/Managers/Commands/CampaignCommands.cs
@@ -113,7 +113,7 @@ namespace WorldServer.Managers.Commands
         public static void LockPairing(Player plr, Realms realm, int forceNumberOfBags=0)
         {
             plr.SendClientMessage($"Attempting to lock the {plr.Region.Campaign.CampaignName} campaign... (call AdvancePairing <realm> <tier> to move ahead)");
-            WorldMgr.GetRegion(plr.Region.RegionId, false).Campaign.BattleFrontManager.LockActiveBattleFront(realm, forceNumberOfBags);
+            WorldMgr.GetRegion(plr.Region.RegionId, false).Campaign.BattleFrontManager.LockActiveBattleFront(realm, (byte) forceNumberOfBags);
         }
 
         [CommandAttribute(EGmLevel.EmpoweredStaff, "Advances the pairing the player is in ")]
